Add GrossTypeLabelsComponent for rendering clickable labels in query results
